feat: finish form with bugs
