+++ src/helpers/index.js
@@ -1,5 +1,4 @@
 export * from './FormValidation';
 export * from './handleUploadFile';
 export * from './fromCategoryToRoute';
-export * from './formValidationNew';
 export * from './requestErrorPopUp';
